// src/components/Forms/Return/tipo-demanda/interfaces.ts
export interface Escritorio {
  officeName: string;
  officeID: number;
  suitID: number;
  fullName: string;
  ResultCodeMsg: string | null;
  userEmail: string;
  userName: string;
  officeEmail: string;
  ResultCode: number;
}

export interface DemandaCadastroJuridicoReturn {
  pk?: string;
  usuario_atuando?: string;
  created_by?: string;
  cliente?: string;
  documento_autor?: string;
  processo?: string;
  data_carimbo?: string;
  data_citacao?: string;
  perfil_demanda?: string;
  status?: string;
  data_distribuicao?: string;
  data_audiencia?: string;
  data_liminar?: string;
  data_citacao_positiva?: string;
  due_date?: string;
  tipo_acao?: string;
  causa_pedir?: string;
  contrato?: ContratoDemanda[];
  prazo_liminar?: string;
  andamento_cedente_assumiu_demanda?: boolean | undefined;
  prazo_contestacao?: string;
  detalhe_objeto?: string;
  arquivos?: ArquivoDemanda[];
  tribunal?: string;
  uf?: string;
  comarca?: string;
  foro?: string;
  vara?: string;
  empresa_fundo?: string;
  tipo_justica?: string;
  valor_causa?: string;
  nome_autor?: string;
  nome_advogado?: string;
  documento_advogado?: string;
  portfolio?: string[];
  escritorio?: string | string[] | Escritorio | Escritorio[];
  tipo_demanda?: string;
  status_demanda?: string;
  observacao?: {
    mensagem: string;
    criada_em: string;
    criada_por: string;
  }[];
}
// ...
export interface ContratoDemanda {
  //Desde 24/02/2025
  bindingId: number; //OK
  ContratoOriginal: string; //OK
  debtId: number; //OK
  portfolioName: string; //OK
  PrimeiroAtraso: string; //OK
  SaldoOriginal?: number; //OK
  selected: boolean; //OK
  located: boolean; //OK

  //Antes de 24/02/2025
  binding_id?: number;
  contrato_original?: string | null;
  debit_id?: number;
  portfolio_name?: string;
  primeiro_atraso?: string;
  saldo_original?: number;

  // //v2 dynamodb
  // bindingId: 42969907,
  // CodigoProduto: "",
  // ContratoOriginal: "C26422463061865600060806639377",
  // debtId: 45952274,
  // DigitoConta: null,
  // located: true,
  // NomeProduto: "CARTÃO MATEUSCARD VISA NACIONAL",
  // NumDivida: "C264224645952274",
  // NumeroAgencia: null,
  // NumeroBanco: null,
  // NumeroCarteira: null,
  // NumeroConta: null,
  // portfolioID: 125,
  // portfolioName: "Bradesco VI (P2)",
  // PrimeiroAtraso: "2020-12-28T00:00:00",
  // SaldoAtualizado: 338.54,
  // SaldoOriginal: 224,
  // selected: true,
  // Situacao: ""

  // //V1 dynamodb
  // binding_id: number;
  // codigo_produto?: string | null;
  // contrato_original?: string | null;
  // debit_id: number;
  // digito_conta?: string | number | null;
  // escritorio?: string;
  // id_portfolio_contrato: number;
  // located: boolean;
  // nome_parte: string;
  // nome_produto?: string | null;
  // num_agencia?: string | number | null;
  // num_banco?: string | number | null;
  // num_carteira?: string | number | null;
  // num_conta?: string | number | null;
  // num_divida?: string | number;
  // portfolio_name: string;
  // primeiro_atraso: string;
  // saldo_atualizado?: number;
  // saldo_original?: number;
  // selected: boolean;
  // situacao?: string;


}